Quarter_03/Class_Challanges/27-Apr-2025/_01/pyproject.toml
[project]
name = "agent-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.13",
]
